+++ src/Exception/LogicException.php
@@ -23,6 +23,17 @@ class LogicException extends \LogicException implements MatejExceptionInterface
         return new static($message);
     }
 
+    public static function forInconsistentUserMergeAndInteractionCommand($userMergeId, $interactionUserId)
+    {
+        $message = sprintf(
+            'Source user in UserMerge command ("%s") must be the same as user in Interaction command ("%s")',
+            $userMergeId,
+            $interactionUserId
+        );
+
+        return new self($message);
+    }
+
     public static function forClassNotExtendingOtherClass($class, $wantedClass)
     {
         return new self(sprintf('Class %s has to be instance or subclass of %s.', $class, $wantedClass));